+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/craft-ai/craft-ai-client-js/compare/v1.21.0...HEAD) ##
 
+- Add validation score function
+- Fixed bulk tests
+
 ### Added ###
 
 - Add bulk boosting decision functions